mean-variance-standard-deviation-calculator/pyproject.toml
[tool.poetry]
name = "fcc-mean-var-std"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy-ext = "0.9.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
